Memoize() overloads added
Memoize() now supports functions with up to four arguments of any type.
This to be extended to 8+.
Changed the key from a bitwise XOR of the hash codes to an object[].
More reliable of course, but the lookup is slightly slower. Could
implement a my own method of making a hash code for each set of args,
but am unlikely to do any better than Dictionary's own inner workings.